refactor: unify course number/range logic
